src/api/requirements.txt
flask
jsonify
python-dotenv
pymongo
